+++ app/components/NavBar/SignInModal.tsx
@@ -1,12 +1,12 @@
 'use client'
 
-import React, { useState, useContext } from 'react'
 import type { FC } from 'react'
+
+import React, { useState, useContext } from 'react'
 import AppContext from '../../../context/AppContext'
 
 import { useSignMessage, useAccount } from 'wagmi'
-
-import EthereumAuth from '../EthereumAuth'
+import EthereumAuth from '../Ethereum/EthereumAuth'
 
 import { Typography, Modal, Col } from 'antd'
