+++ package.json
@@ -16,7 +16,10 @@
 		"axios": "^1.3.4",
 		"eden-sdk": "^0.2.31",
 		"ethers": "<6",
+		"extract-colors": "^3.0.0",
 		"form-data": "^4.0.0",
+		"get-pixels": "^3.3.3",
+		"image-palette-core": "^0.2.2",
 		"iron-session": "latest",
 		"jest-worker": "^29.5.0",
 		"lit-html": "^2.6.1",
@@ -24,6 +27,7 @@
 		"nprogress": "^0.2.0",
 		"react": "latest",
 		"react-blockies": "^1.4.1",
+		"react-color-extractor": "^1.1.2",
 		"react-dom": "latest",
 		"react-icons": "^4.8.0",
 		"react-masonry-css": "^1.0.16",
