+++ styles/CreationCard.module.css
@@ -1,3 +1,24 @@
+@keyframes fadeIn {
+	0% {
+		opacity: 0;
+	}
+	100% {
+		opacity: 1;
+	}
+}
+
+.crImageLinkWrapper {
+	position: relative;
+	width: 100%;
+	overflow: hidden;
+}
+
+.crImageLinkWrapper::before {
+	content: '';
+	display: block;
+	padding-bottom: calc(100% * var(--aspect-ratio));
+}
+
 .crCardWrapper {
 	position: relative;
 	box-shadow: unset !important;
@@ -25,6 +46,7 @@
 .crCard {
 	display: flex;
 	flex-direction: column;
+	animation: fadeIn 0.5s;
 }
 .crLink {
 	display: flex;
